CASE STUDY
Case Study
1. Write a program which will find factors of given number and find whether the
factor is even or odd.
Hint: Use Loop with if-else statements
2. Write a code which accepts a sequence of words as input and prints the words in a
sequence after sorting them alphabetically.
Hint: In case of input data being supplied to the question, it should be assumed to
be a console input.
3. Write a program, which will find all the numbers between 1000 and 3000 (both
included) such that each digit of a number is an even number. The numbers
obtained should be printed in a comma separated sequence on a single line.
Hint: In case of input data being supplied to the question, it should be assumed to
be a console input. Divide each digit with 2 and verify is it even or not.
4. Write a program that accepts a sentence and calculate the number of letters and
digits.
Suppose if the entered string is: Python0325
Then the output will be:
LETTERS: 6
DIGITS:4
Hint: Use built-in functions of string.
5. Design a code which will find the given number is Palindrome number or not.
Hint: Use built-in functions of string.